test(outlierdetection): convert pick_first health listener test to e2e
