+++ internal/xds/balancer/outlierdetection/balancer_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/balancer"
-	"google.golang.org/grpc/balancer/pickfirst"
 	"google.golang.org/grpc/balancer/weightedroundrobin"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/connectivity"
@@ -1103,95 +1102,6 @@ func (s) TestConcurrentOperations(t *testing.T) {
 	od.Close()
 	close(finished)
 	wg.Wait()
-}
-
-// Test verifies that outlier detection doesn't eject subchannels created by
-// the new pickfirst balancer when pickfirst is a non-leaf policy, i.e. not
-// under a petiole policy. When pickfirst is not under a petiole policy, it will
-// not register a health listener. pickfirst will still set the address
-// attribute to disable ejection through the raw connectivity listener. When
-// Outlier Detection processes a health update and sees the health listener is
-// enabled but a health listener is not registered, it will drop the ejection
-// update.
-func (s) TestPickFirstHealthListenerDisabled(t *testing.T) {
-	backend := &stubserver.StubServer{
-		EmptyCallF: func(context.Context, *testpb.Empty) (*testpb.Empty, error) {
-			return nil, errors.New("some error")
-		},
-	}
-	if err := backend.StartServer(); err != nil {
-		t.Fatalf("Failed to start backend: %v", err)
-	}
-	defer backend.Stop()
-	t.Logf("Started bad TestService backend at: %q", backend.Address)
-
-	// The interval is intentionally kept very large, the interval algorithm
-	// will be triggered manually.
-	odCfg := &LBConfig{
-		Interval:         iserviceconfig.Duration(300 * time.Second),
-		BaseEjectionTime: iserviceconfig.Duration(300 * time.Second),
-		MaxEjectionTime:  iserviceconfig.Duration(500 * time.Second),
-		FailurePercentageEjection: &FailurePercentageEjection{
-			Threshold:             50,
-			EnforcementPercentage: 100,
-			MinimumHosts:          0,
-			RequestVolume:         2,
-		},
-		MaxEjectionPercent: 100,
-		ChildPolicy: &iserviceconfig.BalancerConfig{
-			Name: pickfirst.Name,
-		},
-	}
-
-	lbChan := make(chan *outlierDetectionBalancer, 1)
-	bf := stub.BalancerFuncs{
-		Init: func(bd *stub.BalancerData) {
-			bd.ChildBalancer = balancer.Get(Name).Build(bd.ClientConn, bd.BuildOptions)
-			lbChan <- bd.ChildBalancer.(*outlierDetectionBalancer)
-		},
-		Close: func(bd *stub.BalancerData) {
-			bd.ChildBalancer.Close()
-		},
-		UpdateClientConnState: func(bd *stub.BalancerData, ccs balancer.ClientConnState) error {
-			ccs.BalancerConfig = odCfg
-			return bd.ChildBalancer.UpdateClientConnState(ccs)
-		},
-	}
-
-	stub.Register(t.Name(), bf)
-
-	opts := []grpc.DialOption{
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithDefaultServiceConfig(fmt.Sprintf(`{ "loadBalancingConfig": [{%q: {}}] }`, t.Name())),
-	}
-	cc, err := grpc.NewClient(backend.Address, opts...)
-	if err != nil {
-		t.Fatalf("grpc.NewClient() failed: %v", err)
-	}
-	defer cc.Close()
-	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
-	defer cancel()
-	testServiceClient := testgrpc.NewTestServiceClient(cc)
-	testServiceClient.EmptyCall(ctx, &testpb.Empty{})
-	testutils.AwaitState(ctx, t, cc, connectivity.Ready)
-
-	// Failing request should not cause ejection.
-	testServiceClient.EmptyCall(ctx, &testpb.Empty{})
-	testServiceClient.EmptyCall(ctx, &testpb.Empty{})
-	testServiceClient.EmptyCall(ctx, &testpb.Empty{})
-	testServiceClient.EmptyCall(ctx, &testpb.Empty{})
-
-	// Run the interval algorithm.
-	select {
-	case <-ctx.Done():
-		t.Fatal("Timed out waiting for the outlier detection LB policy to be built.")
-	case od := <-lbChan:
-		od.intervalTimerAlgorithm()
-	}
-
-	shortCtx, shortCancel := context.WithTimeout(ctx, defaultTestShortTimeout)
-	defer shortCancel()
-	testutils.AwaitNoStateChange(shortCtx, t, cc, connectivity.Ready)
 }
 
 // Tests handling of endpoints with multiple addresses. The test creates two
